Added PiPiP in Chat pane of CallWithChat composite (mobile view) (#1617)
* Added MobilePaneWithLocalAndRemotePIP

* Refactored EmbeddedPeoplePane

* Added docs and comments

* Change files

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Separate LocalRemotePIP and Modal from MobilePane

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* working checkpoint

* rename and cleanup

* more refactoring. memoized chat.

* more refactoring

* renames

* moved drawer to CallWithChatPane

* rename variables

* fixed dependency useMemo array

* Update composite automation snapshots

* maxWidth style update for desktop pane and maxHeight style update for
copy link button

* modal styles change to type Partial<IModalStyles>

* created type CallWithChatPaneOption

* more readable CallWithChatPane component

* LocalAndRemotePIPInModal renamed to ModalLocalAndRemotePIP

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Added side pane tokens

* Remove export

* Update composite automation snapshots

* Update composite automation snapshots

* fix comment

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Update composite automation snapshots

* Fixed heading text logic

* BasicHeader renamed to SidePaneHeader

* Update composite automation snapshots
