+++ change/@internal-react-composites-d8da4cf2-00e1-48b9-84d0-0e7dfcd0f453.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Added local and remote picture-in-picture component in Chat pane of CallWithChat composite in mobile view.",
+  "packageName": "@internal/react-composites",
+  "email": "miguelgamis@microsoft.com",
+  "dependentChangeType": "patch"
+}
