+++ packages/react-composites/src/composites/CallWithChatComposite/styles/PeoplePaneContent.styles.ts
@@ -1,16 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import {
-  IButtonStyles,
-  IModalStyleProps,
-  IModalStyles,
-  IStackItemStyles,
-  IStackStyles,
-  IStackTokens,
-  IStyleFunctionOrObject,
-  Theme
-} from '@fluentui/react';
+import { IButtonStyles, IStackItemStyles, IStackStyles, IStackTokens } from '@fluentui/react';
 
 /**
  * @private
@@ -57,33 +48,6 @@ export const copyLinkButtonStyles: IButtonStyles = {
  * @private
  */
 export const linkIconStyles = { marginRight: '0.5rem' };
-
-/**
- * @private
- */
-export const modalStyle = (theme: Theme): IStyleFunctionOrObject<IModalStyleProps, IModalStyles> => {
-  return {
-    main: {
-      minWidth: 'min-content',
-      minHeight: 'min-content',
-      position: 'absolute',
-      zIndex: 1,
-      borderRadius: theme.effects.roundedCorner4,
-      boxShadow: theme.effects.elevation8,
-      overflow: 'hidden',
-      ...(theme.rtl ? { left: '1rem' } : { right: '1rem' }),
-      // pointer events for root Modal div set to auto to make LocalAndRemotePIP interactive
-      pointerEvents: 'auto',
-      touchAction: 'none'
-    },
-    root: {
-      width: '100%',
-      height: '100%',
-      // pointer events for root Modal div set to none to make descendants interactive
-      pointerEvents: 'none'
-    }
-  };
-};
 
 /**
  * @private
